+++ Algorithm.CSharp/IndexOptionBuySellCallIntradayRegressionAlgorithm.cs
