+++ Algorithm.CSharp/IndexOptionCallOTMExpiryRegressionAlgorithm.cs
@@ -16,6 +16,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection;
 using QuantConnect.Data;
 using QuantConnect.Interfaces;
 using QuantConnect.Orders;
@@ -24,20 +25,26 @@ using QuantConnect.Securities;
 namespace QuantConnect.Algorithm.CSharp
 {
     /// <summary>
-    /// This regression algorithm tests In The Money (ITM) index option expiry for calls.
+    /// This regression algorithm tests Out of The Money (OTM) index option expiry for calls.
     /// We expect 2 orders from the algorithm, which are:
     ///
-    ///   * Initial entry, buy SPX Call Option (expiring ITM)
-    ///   * Option exercise, settles into cash
+    ///   * Initial entry, buy SPX Call Option (expiring OTM)
+    ///     - contract expires worthless, not exercised, so never opened a position in the underlying
+    ///
+    ///   * Liquidation of worthless SPX call option (expiring OTM)
     ///
     /// Additionally, we test delistings for index options and assert that our
     /// portfolio holdings reflect the orders the algorithm has submitted.
     /// </summary>
-    public class IndexOptionCallITMExpiryRegressionAlgorithm : QCAlgorithm, IRegressionAlgorithmDefinition
+    /// <remarks>
+    /// Total Trades in regression algorithm should be 1, but expiration is counted as a trade.
+    /// See related issue: https://github.com/QuantConnect/Lean/issues/4854
+    /// </remarks>
+    public class IndexOptionCallOTMExpiryRegressionAlgorithm : QCAlgorithm, IRegressionAlgorithmDefinition
     {
         private Symbol _spx;
         private Symbol _spxOption;
-        private Symbol _expectedOptionContract;
+        private Symbol _expectedContract;
 
         public override void Initialize()
         {
@@ -46,17 +53,17 @@ namespace QuantConnect.Algorithm.CSharp
 
             _spx = AddIndex("SPX", Resolution.Minute).Symbol;
 
-            // Select an index option expiring ITM, and adds it to the algorithm.
+            // Select a index option call expiring OTM, and adds it to the algorithm.
             _spxOption = AddIndexOptionContract(OptionChainProvider.GetOptionContractList(_spx, Time)
-                .Where(x => x.ID.StrikePrice <= 3200m && x.ID.OptionRight == OptionRight.Call && x.ID.Date.Year == 2021 && x.ID.Date.Month == 1)
-                .OrderByDescending(x => x.ID.StrikePrice)
+                .Where(x => x.ID.StrikePrice >= 4250m && x.ID.OptionRight == OptionRight.Call && x.ID.Date.Year == 2021 && x.ID.Date.Month == 1)
+                .OrderBy(x => x.ID.StrikePrice)
                 .Take(1)
                 .Single(), Resolution.Minute).Symbol;
 
-            _expectedOptionContract = QuantConnect.Symbol.CreateOption(_spx, Market.USA, OptionStyle.European, OptionRight.Call, 3200m, new DateTime(2021, 1, 15));
-            if (_spxOption != _expectedOptionContract)
+            _expectedContract = QuantConnect.Symbol.CreateOption(_spx, Market.USA, OptionStyle.European, OptionRight.Call, 4250m, new DateTime(2021, 1, 15));
+            if (_spxOption != _expectedContract)
             {
-                throw new Exception($"Contract {_expectedOptionContract} was not found in the chain");
+                throw new Exception($"Contract {_expectedContract} was not found in the chain");
             }
 
             Schedule.On(DateRules.Tomorrow, TimeRules.AfterMarketOpen(_spx, 1), () =>
@@ -104,9 +111,9 @@ namespace QuantConnect.Algorithm.CSharp
             var security = Securities[orderEvent.Symbol];
             if (security.Symbol == _spx)
             {
-                AssertIndexOptionOrderExercise(orderEvent, security, Securities[_expectedOptionContract]);
+                throw new Exception("Invalid state: did not expect a position for the underlying to be opened, since this contract expires OTM");
             }
-            else if (security.Symbol == _expectedOptionContract)
+            if (security.Symbol == _expectedContract)
             {
                 AssertIndexOptionContractOrder(orderEvent, security);
             }
@@ -115,27 +122,7 @@ namespace QuantConnect.Algorithm.CSharp
                 throw new Exception($"Received order event for unknown Symbol: {orderEvent.Symbol}");
             }
 
-            Log($"{Time:yyyy-MM-dd HH:mm:ss} -- {orderEvent.Symbol} :: Price: {Securities[orderEvent.Symbol].Holdings.Price} Qty: {Securities[orderEvent.Symbol].Holdings.Quantity} Direction: {orderEvent.Direction} Msg: {orderEvent.Message}");
-        }
-
-        private void AssertIndexOptionOrderExercise(OrderEvent orderEvent, Security index, Security optionContract)
-        {
-            var expectedLiquidationTimeUtc = new DateTime(2021, 1, 16, 3, 0, 0);
-
-            // No way to detect option exercise orders or any other kind of special orders
-            // other than matching strings, for now.
-            if (orderEvent.Message.Contains("Option Exercise"))
-            {
-                if (orderEvent.FillPrice != 3200m)
-                {
-                    throw new Exception("Option did not exercise at expected strike price (3200)");
-                }
-
-                if (optionContract.Holdings.Quantity != 0)
-                {
-                    throw new Exception($"Exercised option contract, but we have holdings for Option contract {optionContract.Symbol}");
-                }
-            }
+            Log($"{orderEvent}");
         }
 
         private void AssertIndexOptionContractOrder(OrderEvent orderEvent, Security option)
@@ -146,11 +133,15 @@ namespace QuantConnect.Algorithm.CSharp
             }
             if (orderEvent.Direction == OrderDirection.Sell && option.Holdings.Quantity != 0)
             {
-                throw new Exception($"Holdings were found after a filled option exercise");
+                throw new Exception("Holdings were found after a filled option exercise");
             }
-            if (orderEvent.Message.Contains("Exercise") && option.Holdings.Quantity != 0)
+            if (orderEvent.Direction == OrderDirection.Sell && !orderEvent.Message.Contains("OTM"))
             {
-                throw new Exception($"Holdings were found after exercising option contract {option.Symbol}");
+                throw new Exception("Contract did not expire OTM");
+            }
+            if (orderEvent.Message.Contains("Exercise"))
+            {
+                throw new Exception("Exercised option, even though it expires OTM");
             }
         }
 
@@ -181,33 +172,33 @@ namespace QuantConnect.Algorithm.CSharp
         /// </summary>
         public Dictionary<string, string> ExpectedStatistics => new Dictionary<string, string>
         {
-            {"Total Trades", "3"},
-            {"Average Win", "1.22%"},
-            {"Average Loss", "-7.42%"},
-            {"Compounding Annual Return", "-13.222%"},
-            {"Drawdown", "6.300%"},
-            {"Expectancy", "-0.417"},
-            {"Net Profit", "-6.282%"},
-            {"Sharpe Ratio", "-1.345"},
-            {"Probabilistic Sharpe Ratio", "0.005%"},
-            {"Loss Rate", "50%"},
-            {"Win Rate", "50%"},
-            {"Profit-Loss Ratio", "0.17"},
-            {"Alpha", "-0.105"},
-            {"Beta", "-0.003"},
-            {"Annual Standard Deviation", "0.078"},
-            {"Annual Variance", "0.006"},
-            {"Information Ratio", "0.678"},
-            {"Tracking Error", "0.191"},
-            {"Treynor Ratio", "33.18"},
-            {"Total Fees", "$7.40"},
-            {"Estimated Strategy Capacity", "$1300000.00"},
-            {"Fitness Score", "0.008"},
+            {"Total Trades", "2"},
+            {"Average Win", "0%"},
+            {"Average Loss", "-4.03%"},
+            {"Compounding Annual Return", "-8.595%"},
+            {"Drawdown", "4.000%"},
+            {"Expectancy", "-1"},
+            {"Net Profit", "-4.029%"},
+            {"Sharpe Ratio", "-1.294"},
+            {"Probabilistic Sharpe Ratio", "0.017%"},
+            {"Loss Rate", "100%"},
+            {"Win Rate", "0%"},
+            {"Profit-Loss Ratio", "0"},
+            {"Alpha", "-0.069"},
+            {"Beta", "-0.002"},
+            {"Annual Standard Deviation", "0.053"},
+            {"Annual Variance", "0.003"},
+            {"Information Ratio", "0.911"},
+            {"Tracking Error", "0.182"},
+            {"Treynor Ratio", "28.46"},
+            {"Total Fees", "$3.70"},
+            {"Estimated Strategy Capacity", "$830000.00"},
+            {"Fitness Score", "0"},
             {"Kelly Criterion Estimate", "0"},
             {"Kelly Criterion Probability Value", "0"},
-            {"Sortino Ratio", "-0.217"},
-            {"Return Over Maximum Drawdown", "-2.105"},
-            {"Portfolio Turnover", "0.024"},
+            {"Sortino Ratio", "-0.195"},
+            {"Return Over Maximum Drawdown", "-2.134"},
+            {"Portfolio Turnover", "0"},
             {"Total Insights Generated", "0"},
             {"Total Insights Closed", "0"},
             {"Total Insights Analysis Completed", "0"},
@@ -221,7 +212,7 @@ namespace QuantConnect.Algorithm.CSharp
             {"Mean Population Magnitude", "0%"},
             {"Rolling Averaged Population Direction", "0%"},
             {"Rolling Averaged Population Magnitude", "0%"},
-            {"OrderListHash", "c59d790b89d76f1ad3bb7738b28567c9"}
+            {"OrderListHash", "5dc2591837f882d173d2d4852b3b0626"}
         };
     }
 }
