+++ Algorithm.CSharp/IndexOptionPutITMExpiryRegressionAlgorithm.cs
@@ -25,21 +25,16 @@ using QuantConnect.Securities;
 namespace QuantConnect.Algorithm.CSharp
 {
     /// <summary>
-    /// This regression algorithm tests Out of The Money (OTM) index option expiry for puts.
+    /// This regression algorithm tests In The Money (ITM) index option expiry for puts.
     /// We expect 2 orders from the algorithm, which are:
     ///
-    ///   * Initial entry, buy SPX Put Option (expiring OTM)
-    ///     - contract expires worthless, not exercised, so never opened a position in the underlying
-    ///
-    ///   * Liquidation of worthless SPX Put OTM contract
+    ///   * Initial entry, buy ES Put Option (expiring ITM) (buy, qty 1)
+    ///   * Option exercise, receiving cash (sell, qty -1)
     ///
     /// Additionally, we test delistings for index options and assert that our
     /// portfolio holdings reflect the orders the algorithm has submitted.
     /// </summary>
-    /// <remarks>
-    /// Total Trades in regression algorithm should be 1, but expiration is counted as a trade.
-    /// </remarks>
-    public class IndexOptionPutOTMExpiryRegressionAlgorithm : QCAlgorithm, IRegressionAlgorithmDefinition
+    public class IndexOptionPutITMExpiryRegressionAlgorithm : QCAlgorithm, IRegressionAlgorithmDefinition
     {
         private Symbol _spx;
         private Symbol _spxOption;
@@ -54,12 +49,12 @@ namespace QuantConnect.Algorithm.CSharp
 
             // Select a index option expiring ITM, and adds it to the algorithm.
             _spxOption = AddIndexOptionContract(OptionChainProvider.GetOptionContractList(_spx, Time)
-                .Where(x => x.ID.StrikePrice <= 3150m && x.ID.OptionRight == OptionRight.Put && x.ID.Date.Year == 2021 && x.ID.Date.Month == 1)
-                .OrderByDescending(x => x.ID.StrikePrice)
+                .Where(x => x.ID.StrikePrice >= 4200m && x.ID.OptionRight == OptionRight.Put && x.ID.Date.Year == 2021 && x.ID.Date.Month == 1)
+                .OrderBy(x => x.ID.StrikePrice)
                 .Take(1)
                 .Single(), Resolution.Minute).Symbol;
 
-            _expectedContract = QuantConnect.Symbol.CreateOption(_spx, Market.USA, OptionStyle.European, OptionRight.Put, 3150m, new DateTime(2021, 1, 15));
+            _expectedContract = QuantConnect.Symbol.CreateOption(_spx, Market.USA, OptionStyle.European, OptionRight.Put, 4200m, new DateTime(2021, 1, 15));
             if (_spxOption != _expectedContract)
             {
                 throw new Exception($"Contract {_expectedContract} was not found in the chain");
@@ -86,7 +81,7 @@ namespace QuantConnect.Algorithm.CSharp
                 }
                 if (delisting.Type == DelistingType.Delisted)
                 {
-                    if (delisting.Time != new DateTime(2020, 1, 16))
+                    if (delisting.Time != new DateTime(2021, 1, 16))
                     {
                         throw new Exception($"Delisting happened at unexpected date: {delisting.Time}");
                     }
@@ -110,9 +105,9 @@ namespace QuantConnect.Algorithm.CSharp
             var security = Securities[orderEvent.Symbol];
             if (security.Symbol == _spx)
             {
-                throw new Exception("Invalid state: did not expect a position for the underlying to be opened, since this contract expires OTM and is not tradable");
+                AssertIndexOptionOrderExercise(orderEvent, security, Securities[_expectedContract]);
             }
-            if (security.Symbol == _expectedContract)
+            else if (security.Symbol == _expectedContract)
             {
                 AssertIndexOptionContractOrder(orderEvent, security);
             }
@@ -121,7 +116,32 @@ namespace QuantConnect.Algorithm.CSharp
                 throw new Exception($"Received order event for unknown Symbol: {orderEvent.Symbol}");
             }
 
-            Log($"{orderEvent}");
+            Log($"{Time:yyyy-MM-dd HH:mm:ss} -- {orderEvent.Symbol} :: Price: {Securities[orderEvent.Symbol].Holdings.Price} Qty: {Securities[orderEvent.Symbol].Holdings.Quantity} Direction: {orderEvent.Direction} Msg: {orderEvent.Message}");
+        }
+
+        private void AssertIndexOptionOrderExercise(OrderEvent orderEvent, Security index, Security optionContract)
+        {
+            var expectedLiquidationTimeUtc = new DateTime(2021, 1, 15);
+
+            if (orderEvent.Direction == OrderDirection.Buy && orderEvent.UtcTime != expectedLiquidationTimeUtc)
+            {
+                throw new Exception($"Liquidated index option contract, but not at the expected time. Expected: {expectedLiquidationTimeUtc:yyyy-MM-dd HH:mm:ss} - found {orderEvent.UtcTime:yyyy-MM-dd HH:mm:ss}");
+            }
+
+            // No way to detect option exercise orders or any other kind of special orders
+            // other than matching strings, for now.
+            if (orderEvent.Message.Contains("Option Exercise"))
+            {
+                if (orderEvent.FillPrice != 3300m)
+                {
+                    throw new Exception("Option did not exercise at expected strike price (3300)");
+                }
+
+                if (optionContract.Holdings.Quantity != 0)
+                {
+                    throw new Exception($"Exercised option contract, but we have holdings for Option contract {optionContract.Symbol}");
+                }
+            }
         }
 
         private void AssertIndexOptionContractOrder(OrderEvent orderEvent, Security option)
@@ -132,15 +152,11 @@ namespace QuantConnect.Algorithm.CSharp
             }
             if (orderEvent.Direction == OrderDirection.Sell && option.Holdings.Quantity != 0)
             {
-                throw new Exception("Holdings were found after a filled option exercise");
+                throw new Exception($"Holdings were found after a filled option exercise");
             }
-            if (orderEvent.Direction == OrderDirection.Sell && !orderEvent.Message.Contains("OTM"))
+            if (orderEvent.Message.Contains("Exercise") && option.Holdings.Quantity != 0)
             {
-                throw new Exception("Contract did not expire OTM");
-            }
-            if (orderEvent.Message.Contains("Exercise"))
-            {
-                throw new Exception("Exercised option, even though it expires OTM");
+                throw new Exception($"Holdings were found after exercising option contract {option.Symbol}");
             }
         }
 
@@ -171,33 +187,33 @@ namespace QuantConnect.Algorithm.CSharp
         /// </summary>
         public Dictionary<string, string> ExpectedStatistics => new Dictionary<string, string>
         {
-            {"Total Trades", "2"},
-            {"Average Win", "0%"},
-            {"Average Loss", "-5.12%"},
-            {"Compounding Annual Return", "-10.844%"},
-            {"Drawdown", "5.100%"},
-            {"Expectancy", "-1"},
-            {"Net Profit", "-5.116%"},
-            {"Sharpe Ratio", "-1.28"},
-            {"Probabilistic Sharpe Ratio", "0.017%"},
-            {"Loss Rate", "100%"},
-            {"Win Rate", "0%"},
-            {"Profit-Loss Ratio", "0"},
-            {"Alpha", "-0.086"},
-            {"Beta", "-0.003"},
-            {"Annual Standard Deviation", "0.067"},
-            {"Annual Variance", "0.004"},
-            {"Information Ratio", "0.794"},
-            {"Tracking Error", "0.187"},
-            {"Treynor Ratio", "28.078"},
-            {"Total Fees", "$3.70"},
-            {"Estimated Strategy Capacity", "$1200000.00"},
-            {"Fitness Score", "0"},
+            {"Total Trades", "3"},
+            {"Average Win", "4.15%"},
+            {"Average Loss", "-8.27%"},
+            {"Compounding Annual Return", "-9.486%"},
+            {"Drawdown", "4.500%"},
+            {"Expectancy", "-0.249"},
+            {"Net Profit", "-4.457%"},
+            {"Sharpe Ratio", "-1.412"},
+            {"Probabilistic Sharpe Ratio", "0.002%"},
+            {"Loss Rate", "50%"},
+            {"Win Rate", "50%"},
+            {"Profit-Loss Ratio", "0.50"},
+            {"Alpha", "-0.076"},
+            {"Beta", "-0.002"},
+            {"Annual Standard Deviation", "0.053"},
+            {"Annual Variance", "0.003"},
+            {"Information Ratio", "0.871"},
+            {"Tracking Error", "0.183"},
+            {"Treynor Ratio", "37.798"},
+            {"Total Fees", "$7.40"},
+            {"Estimated Strategy Capacity", "$1400000.00"},
+            {"Fitness Score", "0.008"},
             {"Kelly Criterion Estimate", "0"},
             {"Kelly Criterion Probability Value", "0"},
-            {"Sortino Ratio", "-0.193"},
-            {"Return Over Maximum Drawdown", "-2.12"},
-            {"Portfolio Turnover", "0"},
+            {"Sortino Ratio", "-0.238"},
+            {"Return Over Maximum Drawdown", "-2.128"},
+            {"Portfolio Turnover", "0.024"},
             {"Total Insights Generated", "0"},
             {"Total Insights Closed", "0"},
             {"Total Insights Analysis Completed", "0"},
@@ -211,7 +227,7 @@ namespace QuantConnect.Algorithm.CSharp
             {"Mean Population Magnitude", "0%"},
             {"Rolling Averaged Population Direction", "0%"},
             {"Rolling Averaged Population Magnitude", "0%"},
-            {"OrderListHash", "a35054d03fd2caa0a96cbe12e427e928"}
+            {"OrderListHash", "d3fa88c3acadb9345ceac76a2dd3b520"}
         };
     }
 }
